Update project structure for better testing and usability
