+++ Package.swift
@@ -15,12 +15,12 @@ let package = Package(
             name: "UIUnitTest",
             targets: ["UIUnitTest"]
         ),
-        .library(name: "UIUnitTestAPI", targets: ["UIUnitTestAPI"]),
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
         .package(url: "https://github.com/apple/swift-argument-parser", from: "1.3.0"),
+        .package(path: "Server/UIUnitTestAPI"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -43,7 +43,6 @@ let package = Package(
             path: "Lib/Sources/UIUnitTest",
             linkerSettings: [.linkedFramework("XCTest")]
         ),
-        .target(name: "UIUnitTestAPI", path: "Lib/Sources/UIUnitTestAPI"),
     ],
     swiftLanguageModes: [.v6]
 )
